Icon name in .desktop file should not have an extension
This warning:

  WARNING: Icon referenced in desktop file but not exported: com.google.AndroidStudio.png

is caused because the Icon= entry in the .desktop file should not contain the
file extension: the rationale is that an icon is referenced by name, without
specifying the extension, and whoever loads icons can try to load their
preferred format (i.e. there could be also a SVG version).