Allow accessing the home directory and device nodes
The home directory is used for configuration, project files, and downloading
the Android SDK. Access to devices is needed to use the Android Debugging
Bridge (ADB) when the IDE is used to install, run, and debug applications
on an actual device. Unfortunately, Flatpak only allows either accessing only
DRI devices (for graphics), no devices, or all -- so this needs to go for now
with access to all devices, as ADB drives USB devices directly using libusb.