Disable the automatic update checker
The contents of the Flatpak are immutable, so the updater won't be able to
replace the files. This only disables the automatic update checking, and
therefore nag screen asking the user to update. Also, avoid removing the
"build.txt" file, which contains the build number, so Android Studio will
report being version 2.2.2.<buildnumber> instead of just 2.2.2.

https://phabricator.endlessm.com/T14386